added component generator

+++ scripts/generator.js
@@ -1,6 +1,7 @@
 const program = require("commander");
 const fs = require("fs");
 const mkdirp = require("mkdirp");
+const chalk = require("chalk");
 
 const templateIndex = componentName => `
 import ${componentName} from './${componentName}'
@@ -10,29 +11,44 @@ export default ${componentName};
 
 program
   .command("component <component_name>")
+  .alias("c")
   .action(function(component_name, cmd) {
-    mkdirp(`./components/${component_name}`, function(err) {
-      if (err) console.error(err);
+    const componentPath = `./src/components/${component_name}`;
 
-      fs.writeFile("index.js", templateIndex(component_name), function(err) {
-        if (err) throw err;
-        console.log(`created index.js`);
-      });
+    if (fs.existsSync(componentPath)) {
+      console.log(chalk.red(`component ${component_name} already exists`));
+    } else {
+      mkdirp(componentPath, function(err) {
+        if (err) console.error(err);
 
-      fs.writeFile(
-        `${component_name}.js`,
-        templateComponent(component_name),
-        function(err) {
+        fs.writeFile(
+          `${componentPath}/index.js`,
+          templateIndex(component_name),
+          function(err) {
+            if (err) throw err;
+            console.log(chalk.green(`created ${componentPath}/index.js`));
+          }
+        );
+
+        fs.writeFile(
+          `${componentPath}/${component_name}.js`,
+          templateComponent(component_name),
+          function(err) {
+            if (err) throw err;
+            console.log(
+              chalk.green(`created ${componentPath}/${component_name}.js`)
+            );
+          }
+        );
+
+        fs.writeFile(`${componentPath}/Styled.js`, templateStyled(), function(
+          err
+        ) {
           if (err) throw err;
-          console.log(`created ${component_name}.js`);
-        }
-      );
-
-      fs.writeFile("Styled.js", templateStyled(), function(err) {
-        if (err) throw err;
-        console.log(`created index.js`);
+          console.log(chalk.green(`created ${componentPath}/Styled.js`));
+        });
       });
-    });
+    }
   });
 
 program.parse(process.argv);
@@ -46,10 +62,16 @@ export default class ${componentName} extends Component {
     return <div />;
   }
 }
+
+${componentName}.defaultProps = {
+}
 `;
 
 const templateStyled = () => `
 import styled from "styled-components";
 
-export const Foo = styled.div``;
+export const Foo = styled.div\`\`;
+
+Foo.defaultProps = {
+}
 `;
